[FEATURE] #106 Dao 테스트 코드
